handling room clock state when meeting is extended/ended earlier
